refactor: address PR review feedback
Extract PostIdParserTrait for shared CLI helpers.
Add error logging fallback in unwrap_double_paragraphs.
Restore null coalesce in extract_html_blocks.
Simplify ClassicPostFinder mock. Add TODO for
converter stack optimization. Document snippet
truncation magic numbers.
