+++ src/CLI/ConvertCommand.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Apermo\ClassicToGutenberg\CLI;
 
 use Apermo\ClassicToGutenberg\ContentConverter;
-use Apermo\ClassicToGutenberg\Migration\ClassicPostFinder;
 use Apermo\ClassicToGutenberg\Migration\MigrationResult;
 use Apermo\ClassicToGutenberg\Migration\MigrationRunner;
 use Apermo\ClassicToGutenberg\Permission;
@@ -16,6 +15,8 @@ use WP_CLI\Utils;
  * WP-CLI command: classic-to-gutenberg convert.
  */
 class ConvertCommand {
+
+	use PostIdParserTrait;
 
 	/**
 	 * The content converter.
@@ -124,51 +125,6 @@ class ConvertCommand {
 				$this->stats['locked'],
 			),
 		);
-	}
-
-	/**
-	 * Parse post IDs from positional arguments.
-	 *
-	 * Supports space-separated and comma-separated IDs.
-	 *
-	 * @param string[] $args Positional arguments.
-	 *
-	 * @return int[]
-	 */
-	private function parse_post_ids( array $args ): array {
-		if ( $args === [] ) {
-			return [];
-		}
-
-		$post_ids = [];
-		foreach ( $args as $argument ) {
-			foreach ( \explode( ',', $argument ) as $part ) {
-				$part = \trim( $part );
-				if ( $part !== '' && \ctype_digit( $part ) ) {
-					$post_ids[] = (int) $part;
-				}
-			}
-		}
-
-		return $post_ids;
-	}
-
-	/**
-	 * Find classic posts via query.
-	 *
-	 * @param array<string,string> $assoc_args Associative arguments.
-	 *
-	 * @return int[]
-	 */
-	private function find_classic_posts( array $assoc_args ): array {
-		$finder     = new ClassicPostFinder();
-		$query_args = [];
-
-		if ( isset( $assoc_args['post-type'] ) ) {
-			$query_args['post_type'] = \explode( ',', $assoc_args['post-type'] );
-		}
-
-		return $finder->find( $query_args );
 	}
 
 	/**
